+++ lib/compony/components/edit.rb
@@ -33,9 +33,10 @@ module Compony
 
         form_cancancan_action :edit
 
-        action :back_to_owner do
-          next if data_class.owner_model_attr.blank?
-          Compony.button(:show, @data.send(data_class.owner_model_attr), icon: :xmark, color: :secondary, label: I18n.t('compony.cancel'))
+        exposed_intents do
+          if data_class.owner_model_attr
+            add :show, @data.send(data_class.owner_model_attr), icon: :xmark, color: :secondary, label: I18n.t('compony.cancel')
+          end
         end
 
         content :label do