+++ lib/compony/components/form.rb
@@ -25,9 +25,15 @@ module Compony
 
         # Override this to provide a custom submit button
         content :submit_button, hidden: true do
+          # Fake submit button rendered by a button component and submitting the form via JS:
           concat Compony.button_component_class.new(
-            label: @submit_label || I18n.t('compony.components.form.submit'), icon: 'arrow-right', type: :submit
+            label:   @submit_label || I18n.t('compony.components.form.submit'),
+            icon:    'arrow-right', # TODO
+            href:    '#',
+            onclick: "this.closest('form').requestSubmit(); return false;"
           ).render(controller)
+          # Real (but hidden) submit button to allow Return to submit:
+          button type: :submit, hidden: true
         end
 
         # Override this to provide additional submit buttons.