+++ CHANGELOG.md
@@ -1,18 +1,34 @@
 # unreleased
 
-- Implement `Compony::Intent` as discussed in Issue #14:
+- Add `render_sub_comp` helper
+- Major API change: implement `Compony::Intent` as discussed in Issue #14:
     - Rewire many pure helpers to use intents instead, greatly cleaning up their interface
     - Remove pure helpers `rails_action_name`, `path_helper_name`, `comp_cst` and `family_cst`
-    - TODO: `Compony.button`, `compony_link`, guide in the documentation
-- Add `render_sub_comp` helper
+    - Remove `button_defaults` and associated logic, as it was rarely used
+    - Remove `Compony::Components::Button` and replace it by `Compony::Components::Buttons::Link` and `Compony::Components::Buttons::CssButton`
+    - Backport `Compony.button`, `compony_button`, `compony_link`, for backwards-compatibility
+- TODO: Replace Compony.button etc. (currently in backwards-compatible mode), also affects `row_action` in List
+- TODO: documentation
+- TODO: look for further TODOs
 
 ## Steps to take
 
 - Make sure you no longer use the following methods that have been removed:
     - `Compony.rails_action_name`
     - `Compony.path_helper_name`
+    - `Compony.button_defaults`
+    - `Compony.with_button_defaults`
+    - `Compony.button_component_class=` (replace by button styles, see documentation)
     - `Component.comp_cst` (replace by `comp_name`)
     - `Component.family_cst` (replace by `family_name`)
+- If using custom buttons, inherit from `Compony::Components::Buttons::Link` and adjust code as needed. Register your button with `Compony.register_button_style` and consider setting `Compony.default_button_style=`.
+- Search for each following keywords in your application and replace it as follows:
+    - Replace root actions (`action ... do` and `skip_action`)  by exposed intents (see documentation)
+    - Replace `render_actions` or `compony_actions` by a custom loop of the kind `Compony.root_comp&.exposed_intents&.map { |i| i.render(controller) }`
+    - `Compony.button` was typically used in a component's root actions and should thus already have been replaced in the previous step.
+    - Replace `compony_button` by `render_intent`
+      - Replace `label_format` by something like: `render_intent(:show, @data, button: { label: { format: :short } })`
+    - Replace `compony_link` by `render_intent` and pass `button: { style: :link }` as an argument
 
 # 0.7.1